+++ Examples/Schema/Example_SchemaBuilder.ps1
@@ -1,3 +1,5 @@
+Test-Module PSInsight
+
 $JSONpath = '.\PSInsight\Examples\Schema'
 $Schema = Get-Content "$JSONpath\Schema.json" -Raw | ConvertFrom-Json
 Build-Schema -ObjectArray $Schema -InsightApiKey ''